requirements.txt
python-dotenv~=1.0.0
Flask~=2.3.2
aiohttp~=3.9.5
discord.py ~=2.1.0